requirements.txt
line-bot-sdk==1.8.0
flask
firebase-admin
python-dotenv
BrainpyFuck==0.1.1
gunicorn